API

Microbes

MicrobeAgents.AbstractMicrobe — Type

AbstractMicrobe{D} <: AbstractAgent where {D<:Integer}

All microbe types in MicrobeAgents.jl simulations must be instances of user-defined types that are subtypes of AbstractMicrobe. YourMicrobeType{D} <: AbstractMicrobe{D} The parameter D defines the dimensionality of the space in which the microbe type lives (1, 2 and 3 are supported).

All microbe types must have at least the following fields:

id::Int id of the microbe (used internally by Agents.jl)

pos::SVectpr{D,Float64} position of the microbe

vel::SVector{D,Float64} velocity of the microbe

motility::AbstractMotility motile pattern of the microbe

turn_rate::Real average reorientation rate of the microbe

rotational_diffusivity::Real coefficient of brownian rotational diffusion

radius::Real equivalent spherical radius of the microbe

state::Real generic variable for a scalar internal state

MicrobeAgents.Microbe — Type

Microbe{D} <: AbstractMicrobe{D}

Base microbe type for simple simulations.

Microbe has the required fields

id::Int an identifier used internally

pos::SVector{D,Float64} spatial position

vel::SVector{D,Float64} unit velocity vector

speed::Float64 magnitude of the velocity vector

and the default parameters

motility::AbstractMotility = RunTumble() motile pattern of the microbe

turn_rate::Float64 = 1.0 frequency of reorientations

rotational_diffusivity::Float64 = 0.0 coefficient of brownian rotational diffusion

radius::Float64 = 0.0 equivalent spherical radius of the microbe

state::Float64 = 0.0 generic variable for a scalar internal state

MicrobeAgents.BrownBerg — Type

BrownBerg{D} <: AbstractMicrobe{D}

Model of chemotactic E.coli from 'Brown and Berg (1974) PNAS'

Default parameters:

motility = RunTumble()

turn_rate = 1.49 Hz frequency of reorientations

rotational_diffusivity = 0.035 rad²/s coefficient of brownian rotational diffusion

radius = 0.5 μm equivalent spherical radius of the microbe

state = 0.0 corresponds to 'weighted dPb/dt' in the paper

gain = 660 s

receptor_binding_constant = 100 μM

memory = 1 s

MicrobeAgents.Brumley — Type

Brumley{D} <: AbstractMicrobe{D}

Model of chemotactic bacterium from 'Brumley et al. (2019) PNAS'. The model is optimized for simulation of marine bacteria and accounts for the presence of (gaussian) sensing noise in the chemotactic pathway.

Default parameters:

motility = RunReverseFlick(speed = [46.5])

turn_rate = 2.22 Hz → '1/τ₀'

state = 0.0 → 'S'

rotational_diffusivity = 0.035 rad²/s

memory = 1.3 s → 'τₘ'

gain_receptor = 50.0 μM⁻¹ → 'κ'

gain = 50.0 → 'Γ'

chemotactic_precision = 6.0 → 'Π'

radius = 0.5 μm → 'a'

MicrobeAgents.Celani — Type

Celani{D} <: AbstractMicrobe{D}

Model of chemotactic bacterium using the response kernel from 'Celani and Vergassola (2010) PNAS', extracted from experiments on E. coli.

Sensing noise (not present in the original model) is customarily introduced through the molecular counting noise formula by Berg and Purcell, and can be tuned through a chemotactic_precision factor inspired by 'Brumley et al. (2019) PNAS' (defaults to 0, i.e. no noise).

Default parameters:

motility = RunTumble(speed = [30.0])

turn_rate = 1.49 Hz

state = 0

rotational_diffusivity = 0.26 rad²/s

gain = 50.0

memory = 1 s

radius = 0.5 μm

MicrobeAgents.Xie — Type

Xie{D} <: AbstractMicrobe{D}

Model of chemotactic bacterium adapted from 'Xie et al. (2019) Biophys J'. The model is developed based on experimental measurements of the chemotactic response function in the marine bacterium V. alginolyticus. The peculiarity of the model is the presence of distinct parameters for the forward and backward swimming states.

Sensing noise (not present in the original model) is customarily introduced through the molecular counting noise formula by Berg and Purcell, and can be tuned through a chemotactic_precision factor inspired by 'Brumley et al. (2019) PNAS' (defaults to 0, i.e. no noise).

Default parameters:

motility = RunReverseFlick(speed = [46.5])

turn_rate_forward = 2.3 Hz

turn_rate_backward = 1.9 Hz

state = 0.0 s

state_m = 0.0 s

state_z = 0.0 s

rotational_diffusivity = 0.26 rad²/s

adaptation_time_m = 1.29 s

adaptation_time_z = 0.28 s

gain_forward = 2.7 1/s

gain_backward = 1.6 1/s

binding_affinity = 0.39 μM

chemotactic_precision = 0.0

radius = 0.5 μm

Base.position — Function

position(m::AbstractMicrobe)

Return the position of the microbe.

MicrobeAgents.direction — Function

direction(m::AbstractMicrobe)

Return the direction versor of the microbe.

MicrobeAgents.speed — Function

speed(m::AbstractMicrobe)

Return the speed of the microbe.

MicrobeAgents.velocity — Function

velocity(m::AbstractMicrobe)

Return the velocity vector of the microbe (direction times speed).

MicrobeAgents.motilepattern — Function

motilepattern(m::AbstractMicrobe)

Return the motile pattern of the microbe.

MicrobeAgents.turnrate — Function

turnrate(m::AbstractMicrobe)

Return the unbiased turn rate of the microbe.

MicrobeAgents.rotational_diffusivity — Function

rotational_diffusivity(m::AbstractMicrobe)

Return the rotational diffusivity of the microbe.

MicrobeAgents.radius — Function

radius(m::AbstractMicrobe)

Return the radius of the microbe.

MicrobeAgents.state — Function

state(m::AbstractMicrobe)

Return the internal state of the microbe.

Motility

MicrobeAgents.AbstractMotility — Type

AbstractMotility

General abstract interface for motility patterns.

MicrobeAgents.MotilityOneStep — Type

MotilityOneStep

Type for one-step motility patterns (e.g. RunTumble).

A MotilityOneStep has the fields

speed: distribution of microbe speed, new values extracted after each turn

polar: distribution of polar angles

azimuthal: distribution azimuthal angles

For 2-dimensional microbe types, only polar defines reorientations and azimuthal is ignored.

MicrobeAgents.MotilityTwoStep — Type

MotilityTwoStep

Type for two-step motility patterns (e.g. RunReverse, RunReverseFlick) In two-step motility patterns, the two "steps" can have different properties.

A MotilityTwoStep has the fields

speed: distribution of microbe speed, new values extracted after each turn

polar: distribution of in-plane reorientations for motile state fw

azimuthal: distribution of out-of-plane reorientations for motile state fw

speed_backward: distribution of microbe speed, new values extracted after each turn

polar_backward: distribution of in-plane reorientations for motile state bw

azimuthal_backward: distribution of out-of-plane reorientations for motile state bw

motile_state: defines current motile state (e.g. Forward or Backward for a TwoState)

For 2-dimensional microbe types, only polar distributions define reorientations while azimuthal ones are ignored.

MicrobeAgents.RunTumble — Function

RunTumble(; speed=(30.0,), polar=Uniform(-π,π), azimuthal=Arccos())

Constructor for a MotilityOneStep with default values associated to run-and-tumble motion.

MicrobeAgents.RunReverse — Function

RunReverse(; speed=(30.0,), polar=(π,), azimuthal=Arccos(), speed_backward=speed, polar_backward=polar, azimuthal_backward=azimuthal)

Constructor for a MotilityTwoStep with default values associated to run-reverse motion.

MicrobeAgents.RunReverseFlick — Function

RunReverseFlick(; speed=(30.0,), polar=(π,), azimuthal=Arccos(), speed_backward=speed, polar_backward=(-π/2,π/2), azimuthal_backward=azimuthal)

Constructor for a MotilityTwoStep with default values associated to run-reverse-flick motion.

Utils

MicrobeAgents.distance — Function

distance(a, b, model)

Evaluate the euclidean distance between a and b respecting the boundary conditions of the model

MicrobeAgents.distancevector — Function

distancevector(a, b, model)

Evaluate the distance vector from a to b respecting the boundary conditions of the model.

MicrobeAgents.random_speed — Function

random_speed(microbe, model)

Generate a random speed from the motile pattern of microbe.

MicrobeAgents.random_velocity — Function

random_velocity(model)

Generate a random velocity vector with unit norm respecting the dimensionality of model.

MicrobeAgents.Analysis.adf_to_matrix — Function

adf_to_matrix(adf, sym)

Collect quantity sym from the agent dataframe adf and return it in matrix form with dimensions (times, microbes). Requires all microbes to exist from start to end of the simulations (no removals or insertions).

MicrobeAgents.Analysis.adf_to_vectors — Function

adf_to_vectors(adf, sym)

Collect quantity sym from the agent dataframe adf and return a vector of such quantity for each microbe id.

Missing docstring.

Missing docstring for Analysis.unfold. Check Documenter's build log for details.

MeanSquaredDisplacement.unfold! — Function

unfold!(adf, model; key=:position)
unfold!(adf, extent; key=:position)
unfold!(gdf, model; key=:position)
unfold!(gdf, extent; key=:position)

Unfold the periodic trajectories contained in the agent dataframe adf or in the grouped dataframe gdf where each microbe trajectory was grouped by :id. The second argument can be either the spatial extent of the model, or the model itself from which the extent is automatically extracted.

The keyword argument key=:position determines what column of the dataframe has to be unfolded.

Data analysis

MicrobeAgents.Analysis.detect_turns! — Function

detect_turns!(adf; threshold_angle=0.0, kwargs...)

Detect reorientations in the microbe trajectories in adf. Requires adf to have a velocity column. In order to be detected, a reorientation must be larger than the threshold_angle (in radians); the threshold defaults to 0.

A new column is added to the dataframe, with values true when the microbe has turned with respect to the previous frame, or false otherwise.

Keywords

vel_key::Symbol = :velocity: name of the column containing microbe velocities

new_key::Symbol = :has_turned: name of the new column storing turn statistics

MicrobeAgents.Analysis.run_durations — Function

run_durations(adf; turns_key=:has_turned)

Evaluate the duration (in frames) of all runs detected during a simulation. The dataframe should already contain a column of Bools with values true for frames when a turn has occurred and false otherwise (see detect_turns!). The function returns a vector of run durations for each microbe.

turns_key defines the name of the column where the turn statistics is stored (defaults to :has_turned).

MicrobeAgents.Analysis.driftvelocity_point! — Function

driftvelocity_point!(adf, target; normalize=false, kwargs...)
driftvelocity_point!(gdf, target; normalize=false, kwargs...)

Evaluate the drift velocity of microbes towards a target point, extracting their positions and velocities from the agent dataframe adf or from the grouped dataframe gdf (grouped by :id). Requires the existence of columns containing the position and velocity of microbes. By convention the drift velocity will be positive for motion towards the target point.

Keywords

normalize = false: whether drift velocities should be normalized by the instantaneous speed of the microbes

pos_key::Symbol = :position: name of the column containing microbe positions

vel_key::Symbol = :velocity: name of the column containing microbe velocities

MicrobeAgents.Analysis.driftvelocity_direction! — Function

driftvelocity_direction!(adf, target; normalize=false, kwargs...)
driftvelocity_direction!(gdf, target; normalize=false, kwargs...)

Evaluate the drift velocity of microbes along a target direction, extracting their positions and velocities from the agent dataframe adf or from the grouped dataframe gdf (grouped by :id). Requires the existence of a column containing the velocity of microbes. By convention the drift velocity will be positive for motion along the target directoin.

Keywords

normalize = false: whether drift velocities should be normalized by the instantaneous speed of the microbes

vel_key::Symbol = :velocity: name of the column containing microbe velocities

Missing docstring.

Missing docstring for msd. Check Documenter's build log for details.

Missing docstring.

Missing docstring for acf. Check Documenter's build log for details.