First steps

Creating a microbe

Microbes are represented by subtypes of the AbstractMicrobe type, which is in turn a subtype of AbstractAgent introduced by Agents.jl

MicrobeAgents.AbstractMicrobe — Type

AbstractMicrobe{D} <: AbstractAgent where {D<:Integer}

All microbe types in MicrobeAgents.jl simulations must be instances of user-defined types that are subtypes of AbstractMicrobe. YourMicrobeType{D} <: AbstractMicrobe{D} The parameter D defines the dimensionality of the space in which the microbe type lives (1, 2 and 3 are supported).

All microbe types must have at least the following fields:

id::Int id of the microbe (used internally by Agents.jl)

pos::NTuple{D,Float64} position of the microbe

vel::NTuple{D,Float64} velocity of the microbe

motility::AbstractMotility motile pattern of the microbe

turn_rate::Real average reorientation rate of the microbe

rotational_diffusivity::Real coefficient of brownian rotational diffusion

radius::Real equivalent spherical radius of the microbe

state::Real generic variable for a scalar internal state

MicrobeAgents provides different AbstractMicrobe subtypes representing different models of bacterial behavior from the literature.

A basic type, which is typically sufficient for simple motility simulations and does not include chemotaxis, is the Microbe type.

MicrobeAgents.Microbe — Type

Microbe{D} <: AbstractMicrobe{D}

Basic microbe type for simple simulations.

Default parameters:

id::Int = rand(1:typemax(Int)) identifier used internally by Agents.jl

pos::NTuple{D,Float64} = ntuple(zero,D) position

motility = RunTumble() motile pattern

vel::NTuple{D,Float64} = rand_vel(D) unit velocity vector

speed::Float64 = rand_speed(motility) magnitude of velocity vector

turn_rate::Float64 = 1.0 frequency of reorientations

rotational_diffusivity::Real coefficient of brownian rotational diffusion

radius::Float64 = 0.0 equivalent spherical radius of the microbe

state::Float64 = 0.0 generic variable for a scalar internal state

The dimensionality of Microbe must always be specified on creation. All the fields are instead optional, and if not specified will be assigned default values.

To create a Microbe living in a 1-dimensional space, with run-tumble motility and average turn rate $\nu=1\;s^{-1}$, it is therefore sufficient to run

Microbe{1}()

Similarly, for two and three dimensions:

Microbe{2}()
Microbe{3}()

Any custom parameter can be set via kwargs:

Microbe{3}(
    turn_rate = 0.6,
    rotational_diffusivity = 0.1
)

All the other subtypes of AbstractMicrobe work in a similar way, although they will have distinct default values and extra fields.

MicrobeAgents.BrownBerg — Type

BrownBerg{D} <: AbstractMicrobe{D}

Model of chemotactic E.coli from 'Brown and Berg (1974) PNAS'

Default parameters:

motility = RunTumble()

turn_rate = 1.49 Hz frequency of reorientations

rotational_diffusivity = 0.035 rad²/s coefficient of brownian rotational diffusion

radius = 0.5 μm equivalent spherical radius of the microbe

state = 0.0 corresponds to 'weighted dPb/dt' in the paper

gain = 660 s

receptor_binding_constant = 100 μM

memory = 1 s

MicrobeAgents.Brumley — Type

Brumley{D} <: AbstractMicrobe{D}

Model of chemotactic bacterium from 'Brumley et al. (2019) PNAS'. The model is optimized for simulation of marine bacteria and accounts for the presence of (gaussian) sensing noise in the chemotactic pathway.

Default parameters:

motility = RunReverseFlick(speed_forward = [46.5])

turn_rate = 2.22 Hz → '1/τ₀'

state = 0.0 → 'S'

rotational_diffusivity = 0.035 rad²/s

memory = 1.3 s → 'τₘ'

gain_receptor = 50.0 μM⁻¹ → 'κ'

gain = 50.0 → 'Γ'

chemotactic_precision = 6.0 → 'Π'

radius = 0.5 μm → 'a'

MicrobeAgents.Celani — Type

Celani{D} <: AbstractMicrobe{D}

Model of chemotactic bacterium using the response kernel from 'Celani and Vergassola (2010) PNAS', extracted from experiments on E. coli.

Sensing noise (not present in the original model) is customarily introduced through the molecular counting noise formula by Berg and Purcell, and can be tuned through a chemotactic_precision factor inspired by 'Brumley et al. (2019) PNAS' (defaults to 0, i.e. no noise).

Default parameters:

motility = RunTumble(speed = [30.0])

turn_rate = 1.49 Hz

state = 0

rotational_diffusivity = 0.26 rad²/s

gain = 50.0

memory = 1 s

radius = 0.5 μm

MicrobeAgents.Xie — Type

Xie{D} <: AbstractMicrobe{D}

Model of chemotactic bacterium adapted from 'Xie et al. (2019) Biophys J'. The model is developed based on experimental measurements of the chemotactic response function in the marine bacterium V. alginolyticus. The peculiarity of the model is the presence of distinct parameters for the forward and backward swimming states.

Sensing noise (not present in the original model) is customarily introduced through the molecular counting noise formula by Berg and Purcell, and can be tuned through a chemotactic_precision factor inspired by 'Brumley et al. (2019) PNAS' (defaults to 0, i.e. no noise).

Default parameters:

motility = RunReverseFlick(speed_forward = [46.5])

turn_rate_forward = 2.3 Hz

turn_rate_backward = 1.9 Hz

state = 0.0 s

state_m = 0.0 s

state_z = 0.0 s

rotational_diffusivity = 0.26 rad²/s

adaptation_time_m = 1.29 s

adaptation_time_z = 0.28 s

gain_forward = 2.7 1/s

gain_backward = 1.6 1/s

binding_affinity = 0.39 μM

chemotactic_precision = 0.0

radius = 0.5 μm

Creating a model

MicrobeAgents.jl exploits the AgentBasedModel interface from Agents.jl. While the standard Agents.jl syntax will always work, it is typically more convenient to use the method extensions provided by MicrobeAgents.jl, which also includes some default parameters required by the simulations. Both StandardABM and UnremovableABM are supported. Whenever removal of microbes during the simulation is not needed, UnremovableABM is the recommended choice.

Agents.UnremovableABM — Type

UnremovableABM(MicrobeType, space, timestep; kwargs...)

Extension of the Agents.UnremovableABM method for microbe types. Implementation of AgentBasedModel where agents can only be added but not removed. See Agents.AgentBasedModel for detailed information on the keyword arguments.

Arguments

MicrobeType: subtype of AbstractMicrobe{D}, with explicitly specified dimensionality D. A list of available options can be obtained by running subtypes(AbstractMicrobe).

space: a ContinuousSpace{D} with the same dimensionality D as MicrobeType which specifies the spatial properties of the simulation domain.

timestep: the integration timestep of the simulation.

Keywords

properties: additional container of data to specify model-level properties. MicrobeAgents.jl includes a set of default properties (detailed at the end).

scheduler = Schedulers.fastest

rng = Random.default_rng()

spacing = minimum(extent)/20

warn = true

Default properties

When a model is created, a default set of properties is included in the model (MicrobeAgents.default_ABM_properties):

DEFAULT_ABM_PROPERTIES = Dict(
    :t => 0, # counter for timekeeping
    :concentration_field => (pos,model) -> 0.0,
    :concentration_gradient => (pos,model) -> zero.(pos),
    :concentration_time_derivative => (pos,model) -> 0.0,
    # required by models of chemotaxis, default value is glutamate diffusivity
    :compound_diffusivity => 608.0,
    # model stepper, by default only keeps time
    :update! => tick!
)

By including these default properties, we make sure that all the chemotaxis models will work even without extra user intervention. All these properties can be overwritten by simply passing an equivalent key to the properties dictionary when creating the model.

Agents.StandardABM — Type

StandardABM(MicrobeType, space, timestep; kwargs...)

Extension of the Agents.StandardABM method for microbe types. Implementation of AgentBasedModel where agents can be added and removed at any time. If agents removal is not required, it is recommended to use UnremovableABM for better performance. See Agents.AgentBasedModel for detailed information on the keyword arguments.

Arguments

MicrobeType: subtype of AbstractMicrobe{D}, with explicitly specified dimensionality D. A list of available options can be obtained by running subtypes(AbstractMicrobe).

space: a ContinuousSpace{D} with the same dimensionality D as MicrobeType which specifies the spatial properties of the simulation domain.

timestep: the integration timestep of the simulation.

Keywords

properties: additional container of data to specify model-level properties. MicrobeAgents.jl includes a set of default properties (detailed at the end).

scheduler = Schedulers.fastest

rng = Random.default_rng()

warn = true

Default properties

When a model is created, a default set of properties is included in the model (MicrobeAgents.default_ABM_properties):

DEFAULT_ABM_PROPERTIES = Dict(
    :t => 0, # counter for timekeeping
    :concentration_field => (pos,model) -> 0.0,
    :concentration_gradient => (pos,model) -> zero.(pos),
    :concentration_time_derivative => (pos,model) -> 0.0,
    # required by models of chemotaxis, default value is glutamate diffusivity
    :compound_diffusivity => 608.0,
    # model stepper, by default only keeps time
    :update! => tick!
)

By including these default properties, we make sure that all the chemotaxis models will work even without extra user intervention. All these properties can be overwritten by simply passing an equivalent key to the properties dictionary when creating the model.

To create a simple model, we just need to choose a microbe type, the size of the simulation domain and the integration timestep. The properties of the simulation domain are wrapped in the ContinuousSpace object (re-exported from Agents.jl).

extent = (1000.0, 500.0) # size of 2D simulation domain
space = ContinuousSpace(extent)
dt = 0.1 # integration timestep
model = UnremovableABM(Microbe{2}, space, dt)

By default this creates a model with periodic boundary conditions. For hard wall boundary conditions we can instead specify space=ContinuousSpace(extent; periodic=false).

Now bacteria can be added with add_agent! function.

Agents.add_agent! — Function

add_agent!([pos,] [MicrobeType,] model; kwargs...)

MicrobeAgents extension of Agents.add_agent!. Creates and adds a new microbe to model, using the constructor of the agent type of the model. If model accepts mixed agent types, then MicrobeType must be specified. If not specified, pos will be assigned randomly in the model domain.

Keywords can be used to specify default values to pass to the microbe constructor, otherwise default values from the constructor will be used. If unspecified, the function also generates a random velocity vector.

We need not specify anything if we want the microbe to be added at a random position with the default values from the constructor.

add_agent!(model)

The microbe will be now accessible as model[1].

The ABM is now ready to run. We just need to specify how many steps we want to simulate and what data to collect during the run:

nsteps = 100
adf, mdf = run!(model, nsteps; adata=[:pos])

run! will return two dataframes, one for the agent-level data (adf) and one for the model-level data (mdf, which in this case will be empty). This way, we have produced our first random walk. Since adf.pos is a vector of tuples, we first have to unpack the x and y values and then we are ready to plot our trajectory.

using Plots
x = first.(adf.pos)
y = last.(adf.pos)
plot(x, y)

Motility patterns

In MicrobeAgents.jl, motility patterns are represented as instances of AbstractMotility. In particular, currently available patterns are distinguished into two further categories: AbstractMotilityOneStep or AbstractMotilityTwoStep.

MicrobeAgents.AbstractMotilityOneStep — Type

AbstractMotilityOneStep

One-step motility patterns (RunTumble). Subtypes have the following fields:

speed: distribution of microbe speed, new values extracted after each turn

polar: distribution of polar angles

azimuthal: distribution azimuthal angles

For 2-dimensional microbe types, only polar defines reorientations and azimuthal is ignored.

New one-step motility patterns can be created as

MicrobeAgents.@motility NewMotilityType MotilityOneStep AbstractMotilityOneStep begin
    # some extra fields if needed
end

The necessary fields and subtyping will be added automatically, only new extra fields need to be specified in the definition. If default values have to be specified, a constructor needs to be defined explicitly.

MicrobeAgents.AbstractMotilityTwoStep — Type

AbstractMotilityTwoStep

Two-step motility patterns (RunReverse and RunReverseFlick), with different properties between forward and backward state of motion. Subtypes have the following fields:

speed_forward: distribution of microbe speed, new values extracted after each turn

polar_forward: distribution of in-plane reorientations for motile state fw

azimuthal_forward: distribution of out-of-plane reorientations for motile state fw

speed_backward: distribution of microbe speed, new values extracted after each turn

polar_backward: distribution of in-plane reorientations for motile state bw

azimuthal_backward: distribution of out-of-plane reorientations for motile state bw

motile_state: defines current motile state (e.g. Forward or Backward for a TwoState)

For 2-dimensional microbe types, only polar distributions define reorientations while azimuthal ones are ignored.

New two-step motility patterns can be created as

MicrobeAgents.@motility NewMotilityType MotilityTwoStep AbstractMotilityTwoStep begin
    # some extra fields if needed
end

The necessary fields and subtyping will be added automatically, only new extra fields need to be specified in the definition. If default values have to be specified, a constructor needs to be defined explicitly.

One-step motility pattern are characterized by a single swimming stage. Two-step motility patterns instead have two stages which can have distinct properties; these two stages are referred to as "forward" and "backward" but can really represent anything.

MicrobeAgents.jl defines three standard motility patterns:

MicrobeAgents.RunTumble — Method

RunTumble(; speed=(30.0,), polar=Uniform(-π,π), azimuthal=Arccos())

Run-tumble motility. The kwargs speed, polar and azimuthal must be sampleable objects (ranges, arrays, tuples, distributions...), not scalars.

With default values, the reorientations are uniform on the sphere.

MicrobeAgents.RunReverse — Method

RunReverse(;
    speed_forward = (30.0,),
    polar_forward = (π,),
    azimuthal_forward = Arccos(),
    speed_backward = speed_forward,
    polar_backward = polar_forward,
    azimuthal_backward = azimuthal_forward
)

Run-reverse motility, with possibility to have different properties between the forward (run) and backward (reverse) stages. All the fields must be sampleable objects (ranges, arrays, tuples, distributions...), not scalars.

With default values, reorientations are always perfect reversals and the speed is identical between forward and backward runs.

MicrobeAgents.RunReverseFlick — Method

RunReverseFlick(;
    speed_forward = (30.0,),
    polar_forward = (π,),
    azimuthal_forward = Arccos(),
    speed_backward = speed_forward,
    polar_backward = (-π/2, π/2),
    azimuthal_backward = azimuthal_forward
)

Run-reverse-flick motility, with possibility to have different properties between the forward (run) and backward (reverse) stages. All the fields must be sampleable objects (ranges, arrays, tuples, distributions...), not scalars.

With default values, reorientations after forward runs are perfect reversals, while reorientations after backward runs are uniformly distributed on the circle normal to the run direction; speed is identical between forward and backward runs.

The default values provided by the constructors always consider a constant swimming speed of 30 (micron/s) and "ideal" turn-angle distributions. For more accurate simulation where the reorientation statistics of the microbes is important, appropriate distributions should be specified; the constructors will accept any object that can be sampled via rand().