Creating a microbe

Microbes are represented by subtypes of the AbstractMicrobe type, which is in turn a subtype of AbstractAgent introduced by Agents.jl

MicrobeAgents.AbstractMicrobe — Type

AbstractMicrobe{D} <: AbstractAgent where {D<:Integer}

All microbe types in MicrobeAgents.jl simulations must be instances of user-defined types that are subtypes of AbstractMicrobe. YourMicrobeType{D} <: AbstractMicrobe{D} The parameter D defines the dimensionality of the space in which the microbe type lives (1, 2 and 3 are supported).

All microbe types must have at least the following fields:

id::Int id of the microbe (used internally by Agents.jl)

pos::NTuple{D,Float64} position of the microbe

vel::NTuple{D,Float64} velocity of the microbe

motility::AbstractMotility motile pattern of the microbe

turn_rate::Real average reorientation rate of the microbe

rotational_diffusivity::Real coefficient of brownian rotational diffusion

radius::Real equivalent spherical radius of the microbe

state::Real generic variable for a scalar internal state

MicrobeAgents provides different AbstractMicrobe subtypes representing different models of bacterial behavior from the literature.

A basic type, which is typically sufficient for simple motility simulations and does not include chemotaxis, is the Microbe type.

MicrobeAgents.Microbe — Type

Microbe{D} <: AbstractMicrobe{D}

Basic microbe type for simple simulations.

Default parameters:

id::Int = rand(1:typemax(Int)) identifier used internally by Agents.jl

pos::NTuple{D,Float64} = ntuple(zero,D) position

motility = RunTumble() motile pattern

vel::NTuple{D,Float64} = rand_vel(D) unit velocity vector

speed::Float64 = rand_speed(motility) magnitude of velocity vector

turn_rate::Float64 = 1.0 frequency of reorientations

rotational_diffusivity::Real coefficient of brownian rotational diffusion

radius::Float64 = 0.0 equivalent spherical radius of the microbe

state::Float64 = 0.0 generic variable for a scalar internal state

The dimensionality of Microbe must always be specified on creation. All the fields are instead optional, and if not specified will be assigned default values.

To create a Microbe living in a 1-dimensional space, with run-tumble motility and average turn rate $\nu=1\;s^{-1}$, it is therefore sufficient to run

Microbe{1}()

Similarly, for two and three dimensions:

Microbe{2}()
Microbe{3}()

Any custom parameter can be set via kwargs:

Microbe{3}(
    turn_rate = 0.6,
    rotational_diffusivity = 0.1
)

All the other subtypes of AbstractMicrobe work in a similar way, although they will have distinct default values and extra fields.

MicrobeAgents.BrownBerg — Type

BrownBerg{D} <: AbstractMicrobe{D}

Model of chemotactic E.coli from 'Brown and Berg (1974) PNAS'

Default parameters:

motility = RunTumble()

turn_rate = 1.49 Hz frequency of reorientations

rotational_diffusivity = 0.035 rad²/s coefficient of brownian rotational diffusion

radius = 0.5 μm equivalent spherical radius of the microbe

state = 0.0 corresponds to 'weighted dPb/dt' in the paper

gain = 660 s

receptor_binding_constant = 100 μM

memory = 1 s

MicrobeAgents.Brumley — Type

Brumley{D} <: AbstractMicrobe{D}

Model of chemotactic bacterium from 'Brumley et al. (2019) PNAS'. The model is optimized for simulation of marine bacteria and accounts for the presence of (gaussian) sensing noise in the chemotactic pathway.

Default parameters:

motility = RunReverseFlick(speed_forward = [46.5])

turn_rate = 2.22 Hz → '1/τ₀'

state = 0.0 → 'S'

rotational_diffusivity = 0.035 rad²/s

memory = 1.3 s → 'τₘ'

gain_receptor = 50.0 μM⁻¹ → 'κ'

gain = 50.0 → 'Γ'

chemotactic_precision = 6.0 → 'Π'

radius = 0.5 μm → 'a'

MicrobeAgents.Celani — Type

Celani{D} <: AbstractMicrobe{D}

Model of chemotactic bacterium using the response kernel from 'Celani and Vergassola (2010) PNAS', extracted from experiments on E. coli.

Sensing noise (not present in the original model) is customarily introduced through the molecular counting noise formula by Berg and Purcell, and can be tuned through a chemotactic_precision factor inspired by 'Brumley et al. (2019) PNAS' (defaults to 0, i.e. no noise).

Default parameters:

motility = RunTumble(speed = [30.0])

turn_rate = 1.49 Hz

state = 0

rotational_diffusivity = 0.26 rad²/s

gain = 50.0

memory = 1 s

radius = 0.5 μm

MicrobeAgents.Xie — Type

Xie{D} <: AbstractMicrobe{D}

Model of chemotactic bacterium adapted from 'Xie et al. (2019) Biophys J'. The model is developed based on experimental measurements of the chemotactic response function in the marine bacterium V. alginolyticus. The peculiarity of the model is the presence of distinct parameters for the forward and backward swimming states.

Sensing noise (not present in the original model) is customarily introduced through the molecular counting noise formula by Berg and Purcell, and can be tuned through a chemotactic_precision factor inspired by 'Brumley et al. (2019) PNAS' (defaults to 0, i.e. no noise).

Default parameters:

motility = RunReverseFlick(speed_forward = [46.5])

turn_rate_forward = 2.3 Hz

turn_rate_backward = 1.9 Hz

state = 0.0 s

state_m = 0.0 s

state_z = 0.0 s

rotational_diffusivity = 0.26 rad²/s

adaptation_time_m = 1.29 s

adaptation_time_z = 0.28 s

gain_forward = 2.7 1/s

gain_backward = 1.6 1/s

binding_affinity = 0.39 μM

chemotactic_precision = 0.0

radius = 0.5 μm

Creating a model

MicrobeAgents.jl exploits the AgentBasedModel interface from Agents.jl. While the standard Agents.jl syntax will always work, it is typically more convenient to use the method extensions provided by MicrobeAgents.jl, which also includes some default parameters required by the simulations. Both StandardABM and UnremovableABM are supported.

Whenever removal of microbes during the simulation is not needed, UnremovableABM is the recommended choice.