API

Microbes

Base.position — Function

position(m::AbstractMicrobe)

Return the position of the microbe.

MicrobeAgents.direction — Function

direction(m::AbstractMicrobe)

Return the direction versor of the microbe.

MicrobeAgents.speed — Function

speed(m::AbstractMicrobe)

Return the speed of the microbe.

speed(motility::Motility)

Return the speed distribution of the current motile state.

MicrobeAgents.velocity — Function

velocity(m::AbstractMicrobe)

Return the velocity vector of the microbe (direction times speed).

MicrobeAgents.motilepattern — Function

motilepattern(m::AbstractMicrobe)

Return the motile pattern of the microbe.

MicrobeAgents.rotational_diffusivity — Function

rotational_diffusivity(m::AbstractMicrobe)

Return the rotational diffusivity of the microbe.

MicrobeAgents.radius — Function

radius(m::AbstractMicrobe)

Return the radius of the microbe.

MicrobeAgents.state — Function

state(m::AbstractMicrobe)

Return the internal state of the microbe.

state(motility::Motility)

Return the index of active motile state.

MicrobeAgents.motilestate — Function

motilestate(microbe)

Return the current motile state of microbe.

MicrobeAgents.states — Function

state(motility::Motility)

Return the list of all motile states.

MicrobeAgents.transition_weights — Function

transition_weights(motility::Motility, i::Integer)

Return the transition weights from the state with index i towards the other motile states.

MicrobeAgents.duration — Function

duration(motility::Motility)

Return the average unbiased duration of the current motile state.

Missing docstring.

Missing docstring for angle. Check Documenter's build log for details.

MicrobeAgents.distance — Function

distance(a, b, model)

Evaluate the euclidean distance between a and b respecting the boundary conditions of the model

MicrobeAgents.distancevector — Function

distancevector(a, b, model)

Evaluate the distance vector from a to b respecting the boundary conditions of the model.

MicrobeAgents.random_speed — Function

random_speed(microbe, model)

Generate a random speed from the motile pattern of microbe.

MicrobeAgents.random_velocity — Function

random_velocity(model)

Generate a random velocity vector with unit norm respecting the dimensionality of model.

Chemoattractants

MicrobeAgents.AbstractChemoattractant — Type

AbstractChemoattractant{D,T}

Abstract type for chemoattractants. Requires dimensionality (D) and number type (T, typically Float64).

The interface is defined by four core functions that operate on AgentBasedModels:

chemoattractant: returns the chemoattractant object

concentration: returns the function for the concentration field

gradient: returns the function for the concentration gradient

time_derivative: returns the function for the concentration ramp

chemoattractant_diffusivity: returns the thermal diffusivity of the chemoattractant

MicrobeAgents.GenericChemoattractant — Type

GenericChemoattractant{D,T} <: AbstractChemoattractant{D,T}

Type for a generic chemoattractant field. Concentration field, gradient and time derivative default to zero values. Diffusivity defaults to 608 μm^2/s.

MicrobeAgents.chemoattractant — Function

chemoattractant(model)

Returns the chemoattractant object from model.

MicrobeAgents.concentration — Function

concentration(model)

Returns the function f that defines the concentration field. The returned function has signature f(pos, model) and returns a scalar.

MicrobeAgents.gradient — Function

gradient(model)

Returns the function f that defines the gradient of the concentration field. The returned function has signature f(pos, model) and returns a SVector with the same dimensionality as the microbe position pos.

MicrobeAgents.time_derivative — Function

time_derivative(model)

Returns the function f that defines the time derivative of the concentration field. The returned function has signature f(pos, model) and returns a scalar.

MicrobeAgents.chemoattractant_diffusivity — Function

chemoattractant_diffusivity(model)

Returns the thermal diffusivity of the chemoattractant compound.

Utils

MicrobeAgents.Analysis.adf_to_matrix — Function

adf_to_matrix(adf, sym)

Collect quantity sym from the agent dataframe adf and return it in matrix form with dimensions (times, microbes). Requires all microbes to exist from start to end of the simulations (no removals or insertions).

MicrobeAgents.Analysis.adf_to_vectors — Function

adf_to_vectors(adf, sym)

Collect quantity sym from the agent dataframe adf and return a vector of such quantity for each microbe id.

MicrobeAgents.Analysis.unfold — Function

unfold(x::AbstractVector, P)

Unfold timeseries x from a periodic domain extending from 0 to P.

unfold(x1::Real, x0::Real, P::Real)

Unfold the value of x1 with respect to x0 from a domain of periodicity P.

MicrobeAgents.Analysis.unfold! — Function

unfold!(adf, model; key=:position)
unfold!(adf, extent; key=:position)
unfold!(gdf, model; key=:position)
unfold!(gdf, extent; key=:position)

Unfold the periodic trajectories contained in the agent dataframe adf or in the grouped dataframe gdf where each microbe trajectory was grouped by :id. The second argument can be either the spatial extent of the model, or the model itself from which the extent is automatically extracted.

The keyword argument key=:position determines what column of the dataframe has to be unfolded.

unfold!(x::AbstractVector, P)

Unfold timeseries x from a periodic domain extending from 0 to P. The periodicity P can be either a real number (for a cubic domain) or a collection (AbstractVector or NTuple) with one value for each dimension.

Data analysis

MicrobeAgents.Analysis.detect_turns! — Function

detect_turns!(adf; threshold_angle=0.0, kwargs...)

Detect reorientations in the microbe trajectories in adf. Requires adf to have a velocity column. In order to be detected, a reorientation must be larger than the threshold_angle (in radians); the threshold defaults to 0.

A new column is added to the dataframe, with values true when the microbe has turned with respect to the previous frame, or false otherwise.

Keywords

vel_key::Symbol = :velocity: name of the column containing microbe velocities

new_key::Symbol = :has_turned: name of the new column storing turn statistics

MicrobeAgents.Analysis.run_durations — Function

run_durations(adf; turns_key=:has_turned)

Evaluate the duration (in frames) of all runs detected during a simulation. The dataframe should already contain a column of Bools with values true for frames when a turn has occurred and false otherwise (see detect_turns!). The function returns a vector of run durations for each microbe.

turns_key defines the name of the column where the turn statistics is stored (defaults to :has_turned).

MicrobeAgents.Analysis.driftvelocity_point! — Function

driftvelocity_point!(adf, target; normalize=false, kwargs...)
driftvelocity_point!(gdf, target; normalize=false, kwargs...)

Evaluate the drift velocity of microbes towards a target point, extracting their positions and velocities from the agent dataframe adf or from the grouped dataframe gdf (grouped by :id). Requires the existence of columns containing the position and velocity of microbes. By convention the drift velocity will be positive for motion towards the target point.

Keywords

normalize = false: whether drift velocities should be normalized by the instantaneous speed of the microbes

pos_key::Symbol = :position: name of the column containing microbe positions

vel_key::Symbol = :velocity: name of the column containing microbe velocities

MicrobeAgents.Analysis.driftvelocity_direction! — Function

driftvelocity_direction!(adf, target; normalize=false, kwargs...)
driftvelocity_direction!(gdf, target; normalize=false, kwargs...)

Evaluate the drift velocity of microbes along a target direction, extracting their positions and velocities from the agent dataframe adf or from the grouped dataframe gdf (grouped by :id). Requires the existence of a column containing the velocity of microbes. By convention the drift velocity will be positive for motion along the target directoin.

Keywords

normalize = false: whether drift velocities should be normalized by the instantaneous speed of the microbes

vel_key::Symbol = :velocity: name of the column containing microbe velocities

Missing docstring.

Missing docstring for msd. Check Documenter's build log for details.

Missing docstring.

Missing docstring for acf. Check Documenter's build log for details.