API

Data analysis

MicrobeAgents.msd — Function

msd(adf, lags=1:last(adf.step)-1; L=Inf)

Evaluate mean-squared displacement from an agent dataframe adf containing the position timeseries of agents (adf.pos). Parameter L defines the periodicity of the domain for unfolding; set L=Inf (default) if boundaries are not periodic.

msd(x, lags=1:size(x,1)-1)

Evaluate the mean squared displacement of x at timelags lags. If lags is unspecified, all available timepoints are used (lag-0 excluded).

x can be an AbstractVector or an AbstractMatrix. In the latter case, it is assumed that each column represents a distinct signal and each row a timepoint. The msd is then evaluated for each individual timeseries and the results are averaged together.

MicrobeAgents.acf — Function

acf(x, sym, lags=0:last(x.step)-1)

Evaluate the autocorrelation function for quantity sym from dataframe x (x.sym) at timelags lags. If lags is unspecified, the acf is evaluated for all available timepoints.

acf(x, lags=0:size(x,1)-1)

Evaluate the autocorrelation function (acf) of a signal x at timelags lags. If lags is unspecified, the acf is evaluated for all available time points.

The signal x can be an AbstractVector or an AbstractMatrix. In the latter case, it is assumed that each column represents a distinct signal and each row a timepoint. The acf is then evaluated for each individual timeseries and the results are averaged together.

The result is not normalized. If f = acf(x), just evaluate f./f[1] to normalize it.

MicrobeAgents.driftvelocity_direction — Function

driftvelocity_direction(adf, target; normalize=false)

Evaluate the drift velocity of microbes along a direction target, extracting their velocities from the agent dataframe adf (requires :vel field).

Returns a matrix of instantaneous drift velocities with size (nmicrobes, nsteps). By convention the drift velocity will be positive for motion along the target direction.

If normalize is set to true, drift velocities are normalized by the instantaneous speed of the microbe.

MicrobeAgents.driftvelocity_point — Function

driftvelocity_point(adf, target; normalize=false)

Evaluate the drift velocity of microbes towards a point target, extracting their positions and velocities from the agent dataframe adf (requires the existence of :pos and :vel fields). Returns a matrix of instantaneous drift velocities with size (nmicrobes, nsteps). By convention the drift velocity will be positive for motion towards the target point.

If normalize is set to true, drift velocities are normalized by the instantaneous speed of the microbe.

MicrobeAgents.detect_turns — Function

detect_turns(adf; threshold_angle=0.0)

Detect reorientations in the microbe trajectories in adf. Requires adf to have a :vel field containing the unit-norm velocity vector. In order to be detected, a reorientation must be larger than the threshold_angle (in radians); the threshold defaults to 0.

Also works if adf is an AbstractMatrix, with unit velocity vectors for each microbe sorted by column, or if adf is an AbstractVector (unit velocity vectors for a single microbe).

MicrobeAgents.rundurations — Function

rundurations(adf, Δt; threshold_angle=0.0)

Evaluate the duration of all runs observed during a simulation. The dataframe adf should contain the field :vel. Δt is the integration timestep of the simulation. threshold_angle defines the threshold (in radians) to define a reorientation. The function returns a vector of run durations for each microbe.

Also works if adf is an AbstractMatrix, with unit velocity vectors for each microbe sorted by column, or if adf is an AbstractVector (unit velocity vectors for a single microbe).

MicrobeAgents.mean_runduration — Function

mean_runduration(adf, Δt; threshold_angle=0.0)

Evaluate the mean run duration sampled by the simulation. The dataframe adf should contain the field :vel. Δt is the integration timestep of the simulation. threshold_angle defines the threshold (in radians) to define a reorientation.

Also works if adf is an AbstractMatrix, with unit velocity vectors for each microbe sorted by column, or if adf is an AbstractVector (unit velocity vectors for a single microbe).

MicrobeAgents.mean_turnrate — Function

mean_turnrate(adf, Δt; threshold_angle=0.0)

Evaluate the mean turn rate sampled by the simulation. The dataframe adf should contain the field :vel. Δt is the integration timestep of the simulation. threshold_angle defines the threshold (in radians) to define a reorientation.

Also works if adf is an AbstractMatrix, with unit velocity vectors for each microbe sorted by column, or if adf is an AbstractVector (unit velocity vectors for a single microbe).